% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANALYSIS_UTILS.R
\name{block_entropy}
\alias{block_entropy}
\title{compute block entropy}
\usage{
block_entropy(x, translate = TRUE)
}
\arguments{
\item{x}{character matrix or vector of aligned sequences}

\item{translate}{logical. if TRUE (default), translate nucleotide codons
to amino acids before computing entropy}
}
\value{
numeric shannon entropy value (bits, log base 2)
}
\description{
calculates shannon entropy across haplotypes in an msa subset.
optionally translates nucleotide codons to amino acids first.
}
