% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB_UTILS.R
\name{calculate_accessibility}
\alias{calculate_accessibility}
\title{Calculate Solvent Accessibility}
\usage{
calculate_accessibility(
  pdb,
  chain = NULL,
  pdb_path = NULL,
  method = "rose",
  drop_incomplete = F
)
}
\arguments{
\item{pdb}{A bio3d PDB object.}

\item{chain}{Chain ID(s).}

\item{pdb_path}{Path to the PDB file.}

\item{method}{One of \code{'rose'}, \code{'miller'}, \code{'theoretical_tien'}, or \code{'empirical_tien'}.}

\item{drop_incomplete}{Logical, drop residues missing backbone atoms.}
}
\value{
A data frame with residue indices, exposure values, and metadata.
}
\description{
Estimates residue-wise solvent accessibility using a DSSP rewrite.
}
