% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB_UTILS.R
\name{calculate_residue_distance}
\alias{calculate_residue_distance}
\title{Residue-wise Distance Matrix}
\usage{
calculate_residue_distance(
  pdb,
  chain = NULL,
  pdb_path = NULL,
  distance_method = "all"
)
}
\arguments{
\item{pdb}{PDB object or NULL.}

\item{chain}{Chain ID(s).}

\item{pdb_path}{Path to a PDB file.}

\item{distance_method}{One of \code{'all'}, \code{'ca'}, \code{'backbone'}, or \code{'sidechain'}.}
}
\value{
A square numeric matrix of distances.
}
\description{
Computes a pairwise distance matrix between residues based on 3D coordinates.
}
