% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSA_UTILS.R
\name{detect_sequence_type}
\alias{detect_sequence_type}
\title{Detect Sequence Type}
\usage{
detect_sequence_type(ref)
}
\arguments{
\item{ref}{A character string representing a sequence.}
}
\value{
A string: either \code{"nucleotide"} or \code{"protein"}.
}
\description{
Determines whether a given sequence is nucleotide or protein.
}
