% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP_UTILS.R
\name{.plot_coverage}
\alias{.plot_coverage}
\title{Plot Sequence Alignment Coverage}
\usage{
.plot_coverage(coverage)
}
\arguments{
\item{coverage}{A coverage list generated by \code{calculate_coverage()}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Visualizes alignment coverage as horizontal bars for each sequence.
}
