% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB_UTILS.R
\name{download_structures}
\alias{download_structures}
\title{Download PDB Structures}
\usage{
download_structures(hit_table, output_dir = "retrieved_pdbs")
}
\arguments{
\item{hit_table}{Data frame of BLAST hits or character vector of PDB IDs.}

\item{output_dir}{Output directory to save downloaded PDB files.}
}
\value{
A data frame with PDB IDs, chain IDs, and local file paths.
}
\description{
Downloads structures from a table of BLAST hits or PDB IDs.
}
