% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evo3D-package.R
\docType{package}
\name{evo3D-package}
\alias{evo3D}
\alias{evo3D-package}
\title{evo3D: Patch-Level Selection Analysis Using Structure-Informed MSA Subsets}
\description{
Defines sliding windows in 3D space based on protein structure, extracts
corresponding subsets from a codon-aligned MSA, and computes selection statistics
such as nucleotide diversity, Tajima's D, and haplotype diversity.
}
\author{
\strong{Maintainer}: Brad Broyles \email{bbroyle@purdue.edu}

}
\keyword{internal}
