% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXTENSION_UTILS.R
\name{extend_msa}
\alias{extend_msa}
\title{Extend Multi-Chain MSA}
\usage{
extend_msa(msa1, msa2, use_sample_names = T)
}
\arguments{
\item{msa1}{First alignment matrix or named list of matrices.}

\item{msa2}{Second alignment matrix or named list of matrices.}

\item{use_sample_names}{Logical. If TRUE, attempts to match sample rows by name before combining.}
}
\value{
Combined alignment matrix or list of matrices, with each sample’s sequence extended across chains.
}
\description{
Combines two MSAs across chains (or independently aligned segments) either by matching sample names or by row order.
}
