% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB_UTILS.R
\name{identify_patches}
\alias{identify_patches}
\title{Identify Surface Patches}
\usage{
identify_patches(
  dist_mat,
  accessibility_df,
  rsa_cutoff = 0.1,
  sasa_cutoff = NULL,
  dist_cutoff = 15,
  max_patch = NULL,
  only_exposed_in_patch = F
)
}
\arguments{
\item{dist_mat}{A residue-residue distance matrix.}

\item{accessibility_df}{Data frame of residue accessibility values (e.g., from \code{calculate_accessibility}).}

\item{rsa_cutoff}{Minimum RSA for centroid selection (default = 0.1).}

\item{sasa_cutoff}{Optional minimum SASA for centroid selection.}

\item{dist_cutoff}{Maximum distance (in Å) for neighbors.}

\item{max_patch}{Optional maximum number of neighbors per patch.}

\item{only_exposed_in_patch}{If TRUE, restricts patch members to surface residues only.}
}
\value{
Updated \code{accessibility_df} with a \code{patch} column listing neighbors.
}
\description{
Defines residue patches around surface-exposed centroids based on RSA, SASA, and spatial distance.
}
