% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRAPPER_MODULE.R
\name{show_evo3d_defaults}
\alias{show_evo3d_defaults}
\title{show default evo3D controls}
\usage{
show_evo3d_defaults(module_name = NULL)
}
\arguments{
\item{module_name}{character. one of "msa","pdb","aln","stat","output".
if NULL (default) prints all}
}
\value{
none. prints to console
}
\description{
prints the default evo3D control settings for one module
or all modules if none specified
}
