% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSA_UTILS.R
\name{translate_dna}
\alias{translate_dna}
\title{Translate DNA to Amino Acids}
\usage{
translate_dna(ref)
}
\arguments{
\item{ref}{A named character string representing a DNA sequence.}
}
\value{
A named character string representing the translated amino acid sequence.
}
\description{
Converts a DNA reference sequence to a peptide sequence using the standard codon table.
}
