% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANALYSIS_UTILS.R
\name{write_patch_fastas_slow}
\alias{write_patch_fastas_slow}
\title{write patch fasta files}
\usage{
write_patch_fastas_slow(msa_subsets, output_dir = "patch_fastas")
}
\arguments{
\item{msa_subsets}{list of msa matrices (as from \code{extract_msa_subsets()})}

\item{output_dir}{directory to write fasta files (default "patch_fastas")}
}
\description{
writes msa subsets to fasta files, one per patch
}
\keyword{internal}
