% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB_MODULE.R
\name{.calculate_residue_distance}
\alias{.calculate_residue_distance}
\title{Residue-wise Distance Matrix}
\usage{
.calculate_residue_distance(
  pdb,
  chain = NA,
  distance_method = "all",
  in_module = F
)
}
\arguments{
\item{pdb}{PDB object or NULL.}

\item{chain}{Chain ID(s).}

\item{distance_method}{One of \code{'all'}, \code{'ca'}, \code{'backbone'}, or \code{'sidechain'}.}

\item{pdb_path}{Path to a PDB file.}
}
\value{
A square numeric matrix of distances.
}
\description{
Computes a pairwise distance matrix between residues based on 3D coordinates.
}
