% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSA_UTILS.R
\name{.standardize_msa_input}
\alias{.standardize_msa_input}
\title{Standardize MSA Input}
\usage{
.standardize_msa_input(msa_mat, msa_path = NULL)
}
\arguments{
\item{msa_mat}{A matrix or list representing the MSA.}

\item{msa_path}{Optional file path to a FASTA file, used if \code{msa_mat} is not provided.}
}
\value{
A character matrix with rows as sequences and columns as alignment positions.
}
\description{
Ensures that the input multiple sequence alignment (MSA) is formatted as a matrix.
}
