% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB_MODULE.R
\name{.standardize_pdb_input}
\alias{.standardize_pdb_input}
\title{Standardize PDB Input}
\usage{
.standardize_pdb_input(pdb)
}
\arguments{
\item{pdb}{A PDB object, or path to pdb or mmCIF file.}

\item{chain}{Character vector of chain IDs, or \code{"all"}. # REMOVED 5/4}
}
\value{
A trimmed \code{bio3d} PDB object containing only selected chains.
}
\description{
Reads a PDB or CIF file and returns a trimmed structure for specified chains.
}
