% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSA_MODULE.R
\name{.translate_dna_to_protein}
\alias{.translate_dna_to_protein}
\title{Translate DNA to Amino Acids}
\usage{
.translate_dna_to_protein(seq)
}
\arguments{
\item{seq}{A named character string representing a DNA sequence.}
}
\value{
A named character string representing the translated amino acid sequence.
}
\description{
Translates a DNA sequence to its corresponding amino acid sequence using the standard genetic code.
Assumes the input sequence is in-frame and codons are contiguous. Gaps or ambiguous bases may result in 'X' or stop codons.
}
\keyword{internal}
