% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRAPPER_MODULE.R
\name{run_evo3d}
\alias{run_evo3d}
\title{Run evo3D Workflow}
\usage{
run_evo3d(
  msa,
  pdb,
  chain = "auto",
  interface_chain = NA,
  occlusion_chain = NA,
  run_selection = TRUE,
  auto_chain_threshold = 0.2,
  write_patch_fastas = FALSE,
  write_evo3d_df = FALSE,
  output_dir = NULL,
  msa_controls = list(),
  pdb_controls = list(),
  aln_controls = list(),
  stat_controls = list()
)
}
\arguments{
\item{msa}{A matrix, file path, or list of MSAs (character matrices, fasta objects, or file paths).}

\item{pdb}{A \code{bio3d} PDB object or list of such objects.}

\item{chain}{Chain ID(s) to analyze. Can be "auto", a character, vector, or nested list.}

\item{interface_chain}{Chain(s) to include in interface-based patching (optional).}

\item{occlusion_chain}{Chain(s) to include in occlusion masking for RSA (optional).}

\item{run_selection}{Logical; whether to calculate nucleotide diversity, Tajima’s D, etc. (default: \code{TRUE}).}

\item{auto_chain_threshold}{Similarity cutoff used to match MSA peptide to PDB chains (default: \code{0.2}).}

\item{write_patch_fastas}{Logical; whether to save individual patch-level MSA FASTA files.}

\item{write_evo3d_df}{Logical; whether to save the final evo3D dataframe to CSV.}

\item{output_dir}{Directory path for writing outputs if above flags are \code{TRUE}.}

\item{msa_controls}{List of control parameters for MSA preprocessing (e.g. \code{ref_method}, \code{force_seqtype}).}

\item{pdb_controls}{List of control parameters for patch definition (e.g. \code{patch.dist.cutoff}, \code{rsa.method}).}

\item{aln_controls}{List of control parameters for MSA–structure alignment (if applicable).}

\item{stat_controls}{List of control parameters for patch-level statistic calculations (e.g. \code{calc_pi}, \code{calc_tajima}).}
}
\value{
A list containing:
\item{evo3d_df}{Final data frame with codon-level mappings and statistics.}
\item{final_msa_subsets}{Patch-level MSAs used in downstream statistics.}
\item{msa_info_sets}{Reference and peptide information for each input MSA.}
\item{pdb_info_sets}{Structure and patch data for each input PDB.}
\item{aln_info_sets}{Codon-level alignment mappings from MSA to PDB.}
\item{call_info}{Cached input metadata and control parameters.}
}
\description{
Full evo3D wrapper to align MSA(s) and PDB(s), generate 3D-defined patches, codon mappings, and selection statistics.
Handles homomultimers and multi-model structures automatically.
}
