% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRAPPER_MODULE.R
\name{show_evo3d_defaults}
\alias{show_evo3d_defaults}
\title{Show Default evo3D Controls}
\usage{
show_evo3d_defaults(module_name = NULL)
}
\arguments{
\item{module_name}{Optional string: one of "msa", "pdb", "aln", or "stat". If `NULL`, shows all.}
}
\value{
None. Prints output to console.
}
\description{
Print default evo3D module control settings (MSA, PDB, alignment, statistics).
}
